requirements.txt
asgiref==3.7.2
Django~=4.2.5
flake8~=3.9.2
isort==5.12.0
mccabe==0.7.0
Pillow==10.0.0
pycodestyle~=2.7.0
pyflakes~=2.3.1
sqlparse~=0.4.3
typing_extensions==4.7.1

defusedxml~=0.7.1
cffi~=1.15.1
pip~=23.1.2
attrs~=23.1.0
wheel~=0.37.1
cryptography~=41.0.4
tornado~=6.3.1
pytz~=2022.6
tomli~=2.0.1
setuptools~=60.2.0
requests~=2.26.0
djangorestframework~=3.14.0